Make lambda paramters common
